Move bugsnag code to error handler module
